requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==3.0.1
Django==4.2.16
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
et_xmlfile==2.0.0
gunicorn==23.0.0
num2words==0.5.14
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
